requirements.txt
torch
torchvision
transformers
openai-clip
nltk
seaborn
matplotlib